services/beampipeline/requirements.txt
apache-beam[gcp]<=2.67.0
google-cloud-bigtable==2.33.0
pyyaml